Move processing logic from gui/ to core/
